Fix: not implement virtual class when WITH_CUSTOM_PREFIX on

+++ src/logging_custom_prefix_unittest.cc
@@ -849,7 +849,7 @@ struct MyLogger : public base::Logger {
   virtual void Write(bool /* should_flush */,
                      time_t /* timestamp */,
                      const char* message,
-                     int length) {
+                     size_t length) {
     data.append(message, length);
   }
 
@@ -981,7 +981,7 @@ struct RecordDeletionLogger : public base::Logger {
   virtual void Write(bool force_flush,
                      time_t timestamp,
                      const char* message,
-                     int length) {
+                     size_t length) {
     wrapped_logger_->Write(force_flush, timestamp, message, length);
   }
   virtual void Flush() { wrapped_logger_->Flush(); }
